table not rerendeing fixed

+++ frontend/src/pages/DashboardPage.tsx
@@ -8,6 +8,7 @@ import AppTable from '../components/AppTable'
 function DashboardPage() {
   const [newGoal, setNewGoal] = React.useState('');
   const queryClient = useQueryClient();
+  const inputRef = React.useRef(null);
 
 
   const { data } = useQuery(GET_GOALS_API, () => makeGetRequest(GET_GOALS_API));
@@ -24,14 +25,13 @@ function DashboardPage() {
     //@ts-ignore
   const { mutate: deleteMutation } = useMutation((body) => makeDeleteRequest(DELETE_GOAL_API(body.id)), {
     onSuccess: () => {
-      // queryClient.invalidateQueries({ queryKey: [GET_GOALS_API] })
-      alert('done delete')
+      queryClient.invalidateQueries({ queryKey: [GET_GOALS_API] })
     },
     onError: (err) => {
       console.log(err)
     }
   })
-  
+   
   
   const handleSubmit = () => {
     if (newGoal === '') return;
@@ -40,9 +40,12 @@ function DashboardPage() {
   const handleEdit = () => {
   }
 
-  const handleDelete = (rowContent: { _id: string }) => {
-    //@ts-ignore
+  const handleDelete = (rowContent: { _id: string, goalname: string }) => {
+    const isConfirm = window.confirm(`Are you sure you want to delete ${rowContent.goalname}`)
+    if (isConfirm) {
+         //@ts-ignore
      deleteMutation({id: rowContent._id})
+    }
   }
 
   return (
@@ -51,18 +54,24 @@ function DashboardPage() {
       <Box p="1em" marginInline="auto" mt="3em">
         <Flex direction="column" gap="1em" >
           <Text as="h3">Create a Goal</Text>
-          <Input
+            <Input
+            ref={inputRef}
             type="text"
             name="goalname"
             value={newGoal}
             onChange={(e) => setNewGoal(e.target.value)}
             placeholder="Your Goal"
-            size="md"
+              size="md"
+              onKeyDown={(e) => {
+                if (e.key === "Enter" && newGoal.trim() !== "") {
+                    handleSubmit()
+                }
+              }}
           />
           <Button onClick={handleSubmit}>Create</Button>
         </Flex>
       </Box>
-        {data && <AppTable tableData={data} handleEdit={handleEdit} handleDelete={handleDelete} />}
+        {data?.length >= 1 && <AppTable tableData={data} handleEdit={handleEdit} handleDelete={handleDelete} />}
         </Box>
     </Box>
   )
